+++ modules/abstract-eth/package.json
@@ -40,13 +40,18 @@
     "@bitgo/sdk-core": "^8.26.0",
     "@bitgo/statics": "^29.0.0",
     "@bitgo/utxo-lib": "^9.16.0",
-    "bignumber.js": "^9.1.1",
-    "ethers": "^5.1.3",
-    "ethereumjs-util": "7.1.5",
     "@ethereumjs/common": "^2.6.5",
     "@ethereumjs/tx": "^3.3.0",
-    "ethereumjs-abi": "^0.6.5",
+    "bignumber.js": "^9.1.1",
     "bn.js": "^5.2.1",
-    "debug": "^3.1.0"
+    "debug": "^3.1.0",
+    "ethereumjs-abi": "^0.6.5",
+    "ethereumjs-util": "7.1.5",
+    "ethers": "^5.1.3",
+    "lodash": "4.17.21",
+    "keccak": "^3.0.3",
+    "secp256k1": "5.0.0",
+    "@bitgo/sdk-lib-mpc": "^8.15.0",
+    "@metamask/eth-sig-util": "^5.0.2"
   }
 }
