+++ modules/sdk-coin-arbeth/tsconfig.json
@@ -17,9 +17,6 @@
       "path": "../sdk-api"
     },
     {
-      "path": "../sdk-coin-eth"
-    },
-    {
       "path": "../sdk-core"
     },
     {
