+++ modules/sdk-coin-eth/package.json
@@ -42,19 +42,13 @@
   "dependencies": {
     "@bitgo/abstract-eth": "^1.6.0",
     "@bitgo/sdk-core": "^8.26.0",
-    "@bitgo/sdk-lib-mpc": "^8.15.0",
     "@bitgo/statics": "^29.0.0",
     "@bitgo/utxo-lib": "^9.16.0",
-    "@ethereumjs/common": "^2.6.5",
     "@ethereumjs/tx": "^3.3.0",
     "@ethereumjs/util": "8.0.3",
-    "@metamask/eth-sig-util": "^5.0.2",
-    "bignumber.js": "^9.0.0",
-    "bn.js": "^5.2.1",
     "ethereumjs-abi": "^0.6.5",
     "ethereumjs-util": "7.1.5",
     "ethers": "^5.1.3",
-    "keccak": "^3.0.3",
     "lodash": "^4.17.14",
     "secp256k1": "5.0.0",
     "superagent": "^3.8.3"
