+++ modules/sdk-coin-ethw/package.json
@@ -40,11 +40,9 @@
     ]
   },
   "dependencies": {
-    "@bitgo/abstract-eth": "^1.6.0",
     "@bitgo/sdk-coin-eth": "^4.10.0",
     "@bitgo/sdk-core": "^8.26.0",
     "@bitgo/statics": "^29.0.0",
-    "bignumber.js": "^9.0.0",
     "ethereumjs-util": "7.1.5",
     "superagent": "^3.8.3"
   },
