+++ modules/sdk-coin-polygon/package.json
@@ -47,10 +47,7 @@
     "@bitgo/statics": "^29.0.0",
     "@bitgo/utxo-lib": "^9.16.0",
     "@ethereumjs/common": "^2.6.5",
-    "@ethereumjs/tx": "^3.3.0",
-    "bignumber.js": "^9.0.0",
     "ethereumjs-abi": "^0.6.5",
-    "lodash": "^4.17.14",
     "superagent": "^3.8.3"
   },
   "devDependencies": {
