+++ modules/statics/src/map.ts
@@ -34,6 +34,25 @@ export class CoinMap {
     }, new CoinMap());
   }
 
+  static coinNameFromChainId(chainId: number): string {
+    const ethLikeCoinFromChainId: Record<number, string> = {
+      1: 'eth',
+      42: 'teth',
+      5: 'gteth',
+      17000: 'hteth',
+      10001: 'ethw',
+      80001: 'tpolygon',
+      137: 'polygon',
+      56: 'bsc',
+      97: 'tbsc',
+      42161: 'arbeth',
+      421614: 'tarbeth',
+      10: 'opeth',
+      11155420: 'topeth',
+    };
+    return ethLikeCoinFromChainId[chainId];
+  }
+
   /**
    * Override `get` to throw if a coin is missing, instead of returning undefined.
    * It will honor key equivalences in case given key is missing.
